fix/ to show item detail slider
